Разбил на модули, исправил ошибку с другим location внутри offer в модуле data.js
