+++ index.html
@@ -228,6 +228,6 @@
     </div>
   </template>
 
-  <script src="js/main.js" defer></script>
+  <script src="js/main.js" type="module" defer></script>
 </body>
 </html>
